requirements.txt
dj-database-url==0.4.2
dj-static==0.0.6
Django==1.11
django-rest-swagger==2.1.2
django-toolbelt==0.0.1
djangorestframework==3.6.3
gunicorn==19.7.1
MarkupSafe==1.0
Pillow==4.1.1
psycopg2==2.7.1
PyYAML==3.12
static3==0.7.0
django-allauth==0.32.0
autoenv
